Added the PointerManager.
In doing this the event system was refactored a bit. Extracted a base class out of the MyGUIEventTranslator that only deals with the native functions. Made the class as it was before MyGUIWidgetEventTranslator as this can translate events on widgets. Renamed MyGUIEventManager to MyGUIWidgetEventManager as it only works for widget based events. There are some events in MyGUI that work differently than the widget events such as the eventMouseCursorChanged in PointerManager. These changes allow those events to be subscribed to as well.

git-svn-id: https://192.168.3.139:8443/svn/Engine/trunk@841 146ae76c-2b19-ef4c-9a69-67eb2b888ad7